Feed reading complete with caveats
Remove the UUID library as it's built-in now. Should've figured that out sooner.
Add an example feed to reuse instead of hammering the site. I'm too lazy to mock this but it'll get there eventually.
Disable the processing check for now but the typical workflow should be the `built_at` date will change as a new job gets added (I assume) and the `last_processed_at` should typically happen in the future because we really only want to check once or twice per day.

Now that we have a feed to work with on the UI, we can start layering the functionality I envisioned.