Finalizing layout with the site
Remove explicit style which looks to be the same color. We'll likely turn the theme switcher off as I don't want to get into the weeds with this.
Fixed padding to take the full width into account and add rounded corners to each item.
Sticky header on top thanks to z-index explicit usage.
Apply dark coloring to the full page instead of a subsection.

We should be relatively 1:1 now though the font is different.
Rather than spend time on this now, I want to add the extra features I had planned to push this over the top.